extracted interaction_recording. More comprehensions
