+++ servirtium/playback.py
@@ -32,6 +32,7 @@ from http.server import HTTPServer, BaseHTTPRequestHandler
 
 from servirtium.markdown_parser import SimpleMarkdownParser, headers_from, get_markdown_file_strings
 
+
 class MockServiceHttpHandler(BaseHTTPRequestHandler):
     invoking_method = "default_value"
 
@@ -61,11 +62,6 @@ class MockServiceHttpHandler(BaseHTTPRequestHandler):
 
 
 parser = SimpleMarkdownParser()
-
-
-def set_markdown_files(markdown_path):
-    files = get_markdown_file_strings(markdown_path)
-    parser._set_mock_files(files)
 
 
 def start():
